perf: u8 instead of u64 shift

+++ engine/src/move_generator/slider_keys.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Key {
     pub magic: u64,
-    pub shift: u64,
+    pub shift: u8,
     pub offset: usize,
 }
